// src/typings/data/user.d.ts
export enum StatIncreaseType {
	MaxHealth = 'maxHealth',
	MaxEnergy = 'maxEnergy',
	MaxHunger = 'maxHunger',
	MaxThirst = 'maxThirst'
}

export enum CurrentRegionType {
	SleepingDens = 'sleeping dens',
	FoodDen = 'food den',
	MedicineDen = 'medicine den',
	Prairie = 'prairie',
	Ruins = 'ruins',
	Lake = 'lake'
}